fix auth middleware
