+++ src/core/state/reducers.js
@@ -5,7 +5,8 @@ import { reducer as formReducer } from 'redux-form';
 import postReducer from 'scenes/Blog/state';
 import { articleReducer } from 'scenes/Dashboard/scenes';
 import { siteUsersReducer } from 'scenes/Dashboard/scenes/Users';
-import { accountReducer, authReducer } from 'scenes/Account';
+import accountReducer from 'scenes/Account/state/account';
+import authReducer from 'scenes/Account/state/auth';
 import boldrReducer from './boldr';
 
 const reducers = combineReducers({
